Devirtualize RenderElement::setStyle(). <https://webkit.org/b/126065>
setStyle() was only virtual in order to let RenderSVGBlock override
the display type in some cases. Devirtualized it and moved the fixup
logic to StyleResolver::adjustRenderStyle().

This hack had an evil twin in RenderElement::initializeStyle() that
also goes away. FIXME--!

Based on a Blink change by Elliott Sprehn.

Reviewed by Antti Koivisto.


Identifier: 144031@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@160906 268f45cc-cd09-0410-ab3c-d52691b4dbfc