Some links at build.webkit.org/dashboard don't show status messages, and aren't keyboard accessible https://bugs.webkit.org/show_bug.cgi?id=126234
Reviewed by Timothy Hatcher.

Changed elements that are semantically links from <div> to <a>.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/StatusLineView.js:
(StatusLineView): Create <a> or <div> elements conditionally on whether we have a URL.
Don't try simulate a link with CSS/JS, it's not needed any more.
(StatusLineView.prototype._clicked): Removed.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Styles/BuildbotQueueView.css:
Removed text-decoration rules, default ones in Main.css now work for queue elements.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Styles/Main.css:
Simplified cursor rules.
Added a universal rule for links to only show underline on hover, as this is what
we want almost everywhere.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Styles/StatusLineView.css:
(.status-line .bubble): Added text-decoration: none, as this is the one place where
we don't want it even on hover.
(.status-line .label): Force display: block for consistent layout between <a>
and <div>.



Identifier: 144183@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@161092 268f45cc-cd09-0410-ab3c-d52691b4dbfc