[GTK] [CMake] Ensure that the autotools build and the CMake install the same files https://bugs.webkit.org/show_bug.cgi?id=116379
Reviewed by Gustavo Noronha Silva.

.:

* Source/PlatformGTK.cmake: Install the documentation.
* Source/cmake/OptionsCommon.cmake: Add a LIBEXEC variable for WebKit2 executables, and
don't use the default locations on GTK+. We will get them from the CMake-provided GNU installation
variables.
* Source/cmake/OptionsGTK.cmake: Setup the installation variables using the ones provided
by CMake's GNU installation directory support. Fix the definition of DATADIR. Always install
the jsc binary.

Source/JavaScriptCore:

* PlatformGTK.cmake: Install API headers, gir files, and the pkg-config file.

Source/WebCore:

* PlatformGTK.cmake: Reformat some install directives to be consistent with the rest of them.
Install the GObject DOM bindings headers.

Source/WebKit:

* PlatformGTK.cmake: Install API headers, gir files, and the pkg-config file. Move
a couple installed and generated headers to the installed headers list.

Source/WebKit2:

* CMakeLists.txt: Install the plugin process to the new LIBEXEC directory, which
for non-GTK+ platforms is the same as EXEC.
* PlatformGTK.cmake: Install API headers, gir files, and the pkg-config file. Move
a couple installed and generated headers to the installed headers list. Rename the
headers list to be consistent with the WebKit1 build. No longer use a variable for
the name of the plugin process, as it's unnecessary.

Identifier: 144390@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@161325 268f45cc-cd09-0410-ab3c-d52691b4dbfc